requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-environ==0.11.2
daphne==4.1.0
channels==4.0.0
channels-redis==4.2.0
drf-yasg==1.21.7
django-cors-headers==4.3.1
whitenoise==6.6.0
requests==2.31.0